### 2021-07-29 #3549 @whedon
[PRE REVIEW]: cerf: A Python package to evaluate the feasibility and costs of power plant siting for alternate futures
**Submitting author:** @crvernon (<a href="http://orcid.org/0000-0002-3406-6214">Chris Vernon</a>)
**Repository:** <a href="https://github.com/IMMM-SFA/cerf" target ="_blank">https://github.com/IMMM-SFA/cerf</a>
**Version:** v2.0.0
