### 2021-07-29 #3549 @kyleniemeyer
Hi @crvernon, thanks for your submission to JOSS! We do not have an editor available in this area right now, so I'm going to put your submission on our waitlist until someone opens up.

In the meantime, could you resolve that DOI error mentioned above?
