+++ cerf/__init__.py
@@ -7,4 +7,4 @@ from .lmp import *
 from .utils import *
 
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"
