+++ docs/source/cerf.rst
@@ -7,7 +7,6 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   cerf.data
    cerf.tests
 
 Submodules
