+++ _sources/install.rst.txt
@@ -1,6 +1,0 @@
-Installation
-------------
-
-**cerf** can be installed via pip by running the following from a terminal window::
-
-    pip install cerf
