+++ _sources/dependencies.rst.txt
@@ -1,22 +1,0 @@
-Dependencies
-------------
-
-=============   ================
-Dependency      Minimum Version
-=============   ================
-numpy           1.19.4
-pandas          1.1.4
-rasterio        1.2.3
-xarray          0.16.1
-PyYAML          5.4.1
-requests        2.25.1
-joblib          1.0.1
-matplotlib      3.3.3
-seaborn         0.11.1
-whitebox        1.5.1
-fiona           1.8.19
-pyproj          3.0.1
-rtree           0.9.7
-shapely         1.7.1
-geopandas       0.9.0
-=============   ================
