### 2021-08-04 #3549 @fraukewiese
@lisazeyen, @willu47, @fwitte
– would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

The submission under review here is [cerf: A Python package to evaluate the feasibility and costs of power plant siting for alternate futures](https://github.com/IMMM-SFA/cerf)
