### 2021-08-10 #3549 @fwitte
Hi @fraukewiese,
I have some experience in this field, but am much more focused on the engineering thermodynamics and economics of individual power plants. Therefore, I feel, that I would not be an adequate reviewer for this publication. I suggest you reach out to e.g. `@simnh` as potential reviewer instead, as he did somewhat related research in the German/European power system. He might be interested even though he is not (yet) on the list of potential reviewers. But thank you for reaching out.
Have a nice day
