### 2021-08-10 #3549 @fraukewiese
@willu47 : Did you find the chance to have a look whether you could be a reviewer for this submission to JOSS?

The submission under review here is cerf: A Python package to evaluate the feasibility and costs of power plant siting for alternate futures
