### 2021-08-11 #3549 @willu47
@fraukewiese 👍🏻 - yes, I am able to perform this review and I am a suitable reviewer - researching spatial electrification pathways, energy-water nexus and multi-scale modelling.

If you need other reviewers, I would recommend any of the following: czor847, ElcoK, tomalrussell (untagged)
