### 2021-08-11 #3601 @crvernon
@fraukewiese @lisazeyen @willu47  The following is a preemptive explanation of author contributions to this paper:

- Chris R. Vernon.  Lead author and developer.
- Jennie S. Rice.  Original product conception, algorithm development, and documentation.
- Nino Zuljevic.  Algorithm, supporting data, and code development.
- Kendall Mongird.  Algorithm, supporting data, and documentation development.
- Kristian Nelson.  Supporting data and documentation development.
- Gokul Iyer.  Supporting data and documentation development.
- Nathalie Voisin.  Supporting data and documentation development.
- Matthew Binsted.  Supporting data and documentation development.
