### 2021-08-11 IMMM-SFA/cerf#37 @willu47
JOSS Review: Clarify license text
Part of openjournals/joss-reviews#3601

Could you clarify that the license text is identical to the [OSI approved 2-clause BSD](https://opensource.org/licenses/BSD-2-Clause)?

- [x] The text formatting of the clauses appears to differ
- [x] the SPDX short identifier differs from `BSD-2-Clause`
