### 2021-08-11 IMMM-SFA/cerf#37 @willu47
I also note that an earlier version (v1.0.0) of cerf appears to contain GPL licensed C code, which may be incompatible with the BSD license. It may be worth checking this.
