### 2021-08-11 IMMM-SFA/cerf#39 @willu47
JOSS Review: Installation
Part of openjournals/joss-reviews#3601

I used miniconda to create a new Python 3.9 environment on OSX 10.15.7 and then ran the command `pip install cerf`. This installed all dependencies without a problem.

- [x] Improve readme for the quickstart tutorial. Dependencies (e.g. Jupyter) for the tutorial notebooks are not installed, and the quickstart tutorial is not available when installing the cerf package (e.g. it is not placed anywhere on the user computer). This would be confusing for a beginner user.  It might be simpler to integrate the information in the notebook into the documentation, and make the notebook downloadable from the docs (with separate installation instructions for the tutorial there).
