### 2021-08-11 IMMM-SFA/cerf#37 @crvernon
Thanks @willu47 .  I had our legal look into the inclusion of GPL products without replication when setting up the original license in 2018.  I will review that documentation and provide clarity.  Also, I will modify the license text to be identical with the OSI approved text and use the correct SPDX short id in the setup args.
