### 2021-08-11 IMMM-SFA/cerf#40 @willu47
JOSS Review: Tests
Part of openjournals/joss-reviews#3601

I cloned the repository to my laptop (OSX 10.15.7) and created a new Python 3.9 environment using miniconda.  I ran `python setup.py develop` to install the package in editable mode.

I manually installed `pytest` and `pytest-cov` and ran the tests using `pytest`. 19 tests passed.
```bash
% pytest
================================================== test session starts ==================================================
platform darwin -- Python 3.9.6, pytest-6.2.4, py-1.10.0, pluggy-0.13.1
rootdir: /Users/wusher/repository/cerf
plugins: cov-2.12.1
collected 19 items

cerf/tests/test_compete.py .                                                                                      [  5%]
cerf/tests/test_lmp.py .                                                                                          [ 10%]
cerf/tests/test_nov.py ...                                                                                        [ 26%]
cerf/tests/test_package_data.py ...........                                                                       [ 84%]
cerf/tests/test_read_config.py .                                                                                  [ 89%]
cerf/tests/test_spatial_reader.py .                                                                               [ 94%]
cerf/tests/test_utils.py .                                                                                        [100%]

================================================== 19 passed in 5.25s ===================================================
```
- [x] Please document in contributing guidelines how to run the tests
- [x] I would recommend pulling the tests out from the `cerf/cerf` subfolder in the repository. These are not needed for the installation of the package.  Also, pulling them out of the package means that a `python setup.py develop` installation is required to run the tests - useful to check the interface to the library.
- [x] `pytest` and `pytest-cov` are not included in develop dependencies (`dev` under `extras_require` in `setup.py`
