### 2021-08-11 IMMM-SFA/cerf#42 @willu47
JOSS Review: Quickstart
Part of openjournals/joss-reviews#3601

I cloned the repository to my laptop (OSX 10.15.7) and created a new Python 3.9 environment using miniconda. I ran python setup.py develop to install the package in editable mode.

I manually installed the `jupyter` and `notebook` packages using miniconda.

I ran all cells in the `quickstart.ipynb` notebook and I get an error running cell 4:

```python
Processing year:  2010
NameError                                 Traceback (most recent call last)
/var/folders/_7/bccbfn110b501smf76mty7xh0000gn/T/ipykernel_22869/3443521374.py in <module>
     11     # do not intialize the run with previously sited data if it is the first time step
     12     if index == 0:
---> 13         result_df = cerf.run(config_file, write_output=False, config_dict=config_dict)
     14
     15     else:

NameError: name 'config_dict' is not defined
```
