### 2021-08-11 IMMM-SFA/cerf#39 @willu47
Hi @crvernon - I think for all these review issues, viewing the final merged product is easiest. Do reference the issue in your PR so I can follow along. I won't have time to review every PR and would prefer you do that in-house.
