### 2021-08-11 IMMM-SFA/cerf#39 @crvernon
@willu47 Sounds great to me.  I'll keep them rolling.  I appreciate the prompt and valuable feedback.
