### 2021-08-11 IMMM-SFA/cerf#43 @willu47
JOSS Review: Requirements
Part of openjournals/joss-reviews#3601

I've noticed in cerf that there are requirements pinned in `requirements.txt` and in `setup.py`. You may want to consider removing the `=~` pins as they're restrictive - you may miss patches in your dependencies. GIven that the test coverage is reasonable this may not be a risk.
