### 2021-08-11 eb03517
update the quickstarter protocol in docs
