### 2021-08-12 IMMM-SFA/cerf#46 @willu47
JOSS Review: Documentation
Part of openjournals/joss-reviews#3601

The documentation is of high quality, readable and holds useful information relevant to new and experienced users of the package.

- There is a clear statement of need
- There are basic installation instructions (use pip) with a list of dependencies
- A user guide provides two example uses of the package in a Jupyter notebook format.
- An API reference provides documentation of the methods available to the user
- There are instructions to potential contributors, but these are currently basic e.g. do not provide information on how to run the tests, or set up a development environment

# Recommendations

## Contributing
- [x] As with #40, document in [contributing.rst](docs/source/contributing.rst) how to set up a development environment, dev dependencies, run the tests, build the documentation etc.
- [x] Consider setting up issue templates on Github [optional]
- [x] Document the release process
- [x] Document continuous integration

## Getting started

- [x] Reorganise quickstart (ticked by #39)
- [x] Clean up hyperlinks - there are a number of links to the notebook contained in the repo e.g. from the "useful links" on the documentation landing page. Please remove extraneous links, or update them to point to the quickstart in the docs.
- [x] Typo: https://immm-sfa.github.io/cerf/user_guide.html#configration-file-setup should be "configuration"
- [x] It is unclear from the [about](https://immm-sfa.github.io/cerf/getting_started.html#about) text how generalisable this package is to other jurisdictions. For example, if I want to run this for an African country, what would I need to do? There's probably a large enough "market" in the USA to justify this type of scientific software, but clarifying this would be informative for potential users.  In addition, understanding what type of models can link to this software would be useful. For example, could I soft-link the outputs from a regionally aggregate energy system model to cerf to gain extra spatial insight?
- [x] Document installation for users with Python environment managers such as venv and conda/miniconda. You could provide an `environment.yaml` download for conda users.

## User guide

A readable and accessible guide. A large part of this section is devoted to writing a configuration file, and describing the data inputs required to run the software. The software is very data intensive, and would provide a barrier to using the software outside of the USA. I wonder if the authors could do anything to make it easier for new users of the software?

- [x] Typo: "The following are the required key, values if your wish to construct your own configuration files:" should read "...required key values if you wish to..."
- [x] [Tutorials section](https://immm-sfa.github.io/cerf/user_guide.html#tutorials) - update to link to quickstart (#39)

## API Reference

The API reference is comprehensive. Docstrings are provided for all classes, methods and functions.

- [x] [Page](https://immm-sfa.github.io/cerf/cerf.html) title not rendered -
- [x] Tests are visible in API docs as a sub-package - should be fixed by #40

- [x] Authors of the documentation are not visible or attributed. Suggest adding from list in [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/3601#issuecomment-896796891).
- [x] Provide hyperlinks/doi to papers which cite or use the software (it's useful for existing or potential users to see papers which use the software)
- [x] Add "how to cite" section to the readme and documentation bibtex or alternatives for JOSS paper

## Other observations

- Using `state_name` in `expansion_plan` key of the configuration file seems to limit generalisability of the tool. Suggest using `region_id` or alternative in future refactors.
