### 2021-08-13 8966214
add pytest and cov to dev extras
