+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,33 @@
+---
+
+name: Bug Report
+about: Create a bug report to help us improve cerf
+title: "BUG:"
+labels: "Bug, Needs Triage"
+
+---
+
+- [ ] I have checked that this issue has not already been reported.
+
+- [ ] I have confirmed this bug exists on the latest version of cerf.
+
+- [ ] (optional) I have confirmed this bug exists on the master branch of cerf.
+
+---
+
+**Note**: Please read [this guide](https://matthewrocklin.com/blog/work/2018/02/28/minimal-bug-reports) detailing how to provide the necessary information for us to reproduce your bug.
+
+#### Code Sample, a copy-pastable example
+
+```python
+# Your code here
+
+```
+
+#### Problem description
+
+[this should explain **why** the current behaviour is a problem and why the expected output is a better solution]
+
+#### Expected Output
+
+[paste your formatted output here]
