+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,33 @@
+---
+
+name: Feature Request
+about: Suggest an idea for cerf
+title: "ENH:"
+labels: "Enhancement, Needs Triage"
+
+---
+
+#### Is your feature request related to a problem?
+
+[this should provide a description of what the problem is, e.g. "I wish I could use cerf to do [...]"]
+
+#### Describe the solution you'd like
+
+[this should provide a description of the feature request]
+
+#### API breaking implications
+
+[this should provide a description of how this feature will affect the API]
+
+#### Describe alternatives you've considered
+
+[this should provide a description of any alternative solutions or features you've considered]
+
+#### Additional context
+
+[add any other context, code examples, or references to existing implementations about the feature request here]
+
+```python
+# Your code here, if applicable
+
+```
