+++ .github/ISSUE_TEMPLATE/documentation_improvement.md
@@ -1,0 +1,20 @@
+---
+
+name: Documentation Improvement
+about: Report wrong or missing documentation
+title: "DOC:"
+labels: "Docs, Needs Triage"
+
+---
+
+#### Location of the documentation
+
+[A link to the file containing the documentation in question]
+
+#### Documentation problem
+
+[this should provide a description of what documentation you believe needs to be fixed/improved]
+
+#### Suggested fix for documentation
+
+[this should explain the suggested fix and **why** it's better than the existing documentation]
