### 2021-08-16 7e65eec
update citations for publications using cerf
