### 2021-08-16 a8a5266
add how to cite to docs
