+++ MANIFEST.in
@@ -1,3 +1,3 @@
-include src/cerf/data/*
+include cerf/data/*
 include tests/data/*
 include tests/data/comp_data/*
