+++ _sources/publications.rst.txt
@@ -1,0 +1,10 @@
+Publications
+============
+
+**cerf v2:**
+
+    Vernon, C.R., Rice, J.S., Zuljevic, N., Mongird, K., Nelson, K., Iyer, G., Voisin, N., Binsted, M. (in review). cerf: A Python package to evaluate the feasibility and costs of power plant siting for alternate futures. Journal of Open Source Software, 6(63), 3549.
+
+**cerf v1:**
+
+    Vernon, C.R., Zuljevic, N., Rice, J.S., Seiple, T.E., Kintner-Meyer, M.C.W., Voisin, N., Kraucunas, I.P., Chunlian, J., Olson, J., Schmidt, L., Morris, S.L. and Patel, P., 2018. CERF – A Geospatial Model for Assessing Future Energy Production Technology Expansion Feasibility. Journal of Open Research Software, 6(1), p.20. DOI: http://doi.org/10.5334/jors.227
