+++ _sources/how_to_cite.rst.txt
@@ -1,0 +1,11 @@
+How to cite **cerf**
+====================
+
+Current software paper status:
+
+.. image:: https://joss.theoj.org/papers/28fee3407bbbef020fb4bb19bd451407/status.svg
+    :target: https://joss.theoj.org/papers/28fee3407bbbef020fb4bb19bd451407
+
+Vernon, C.R., Rice, J.S., Zuljevic, N., Mongird, K., Nelson, K., Iyer, G., Voisin, N., Binsted, M. (in review). cerf: A Python package to evaluate the feasibility and costs of power plant siting for alternate futures. Journal of Open Source Software, 6(63), 3549.
+
+BibTeX citation will be provided once the paper has been accepted.
