+++ _sources/quickstarter2.rst.txt
@@ -1,0 +1,245 @@
+Welcome to the cerf quickstarter!
+=================================
+
+``cerf`` is an open-source geospatial Python package for evaluating and analyzing future electricity technology capacity expansion feasibility.
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Purpose
+-------
+
+``cerf`` was created to:
+
+-  Evaluate the feasibility of a future scenario-driven electricity
+   technology capacity expansion plan as generated by a parent model,
+
+-  Site power plants in the least cost configuration when considering
+   regional economics an on-the-ground barriers to siting,
+
+-  Assist planners and modelers of alternate future realizations of the
+   electricity system to gain an understanding of how siting costs and
+   service area congestion may respond under certain stressors.
+
+A brief introduction
+--------------------
+
+The Capacity Expansion Regional Feasibility package (``cerf``) helps us
+evaluate the feasibility and structure of future electricity capacity
+expansion plans by siting power plants in areas that have been deemed
+the least cost option. We can use ``cerf`` to gain an understanding of
+topics such as: 1) whether or not future projected electricity expansion
+plans from models such as GCAM are possible to achieve, 2) where
+suitability (e.g., cooling water availability) may influence our ability
+to achieve certain expansions, and/or 3) how power plant infrastructure
+build outs and value may evolve into the future when considering
+locational marginal pricing from a grid operations model.
+
+``cerf`` currently operates at a 1 km2 resolution over the conterminous
+United States. Each grid cell is given an initial value of suitable (0)
+or unsuitable (1) based on a collection of suitability criteria gleaned
+from the literature. ``cerf``\ ’s default suitability layers include
+both those that are common to all thermal technologies as well as
+technology-specific suitability criteria. Common suitability layers
+represent categories such as protected lands, critical habitat areas,
+and much more. Technology-specific suitability layers are those that
+satisfy requirements that may not be applicable to all technologies. An
+example would be minimum mean annual flow requirements for cooling water
+availability for individual thermal technologies.
+
+We introduce a metric named Net Locational Cost (NLC) that is used
+compete power plant technologies for each grid cell based on the least
+cost option to site. NLC is calculated by subtracting the Net Operating
+Value (NOV) of a proposed power plant from the cost of its
+interconnection to the grid to represent the potential deployment value.
+Both the NOV parameter which incorporates many technology-specific
+values such as variable operations and maintenance costs, carbon price,
+heat rate, etc. and the interconnection cost parameter used for both
+electricity transmission and gas pipelines are configurable per time
+step. All equations used in ``cerf`` are described in detail in the
[…]

[…]
