### 2021-08-16 IMMM-SFA/cerf#46 @crvernon
Excellent suggestion about the next refactor using something like `region_id` that would be more general.  I did add in a paragraph in the `Getting Started` section of the docs on using **cerf** outside of the context of the US.  Thanks!
