### 2021-08-16 #3601 @crvernon
@willu47 I believe I have incorporated all of your feedback (see PR chain above).  I really appreciate you taking time to make suggestions and edits that were very beneficial to this software!  Please let me know if you have any more questions and I'll be happy to address them.
