### 2021-08-17 IMMM-SFA/cerf#52 @willu47
JOSS REVIEW: Location of data
Part of openjournals/joss-reviews#3601

Hi @crvernon. Many thanks for your quick work on this review.  The package is shaping up nicely.

Currently there is a large (200MB) data folder nestled in the package source. This data is copied over to local machines when users install the package, and it bloats your git repository and the package, wasting bandwidth on PyPI and Github.

On the plus side, it makes your work easily reproducible for your particular analysis - you install the package and run the command.

I would advise removing the data from the package, and hosting it elsewhere, perhaps on Zenodo as a dataset, in the Harvard dataverse or other online data deposit. A helper function (perhaps the code currently in [`package_data.py`](https://github.com/IMMM-SFA/cerf/blob/main/cerf/package_data.py)?) could then retrieve the data (which might be useful for the example notebook), and you would then avoid the repository bloat. This would then be consistent with principle of separation of data and code, where data is specific to a particular analysis; while code is used across analyses.

Alternatively, and if you wish to maximise reproducibility of the analysis for CONUS, deposit the data as advised above, and then create a second Github repo containing a script which installs cerf, downloads the data and runs the analysis.

The focus of this review is the cerf software package, so there's no hard requirement on what you do with the data, although I'd recommend the above. To summarise the steps:

- [x] remove the data from the repository
- [x] ensure that the data (or alternative) is available for the tutorial notebook
