+++ cerf/data/costs_gas_pipeline.yml
@@ -1,2 +1,0 @@
-# cost per kilometer of gas pipeline in $/km
-gas_pipeline_cost: 737000
