+++ cerf/data/state-abbrev_to_state-name.yml
@@ -1,50 +1,0 @@
-# State abbreviation to state name
-AL: alabama
-AZ: arizona
-AR: arkansas
-CA: california
-CO: colorado
-CT: connecticut
-DE: delaware
-DC: district_of_columbia
-FL: florida
-GA: georgia
-ID: idaho
-IL: illinois
-IN: indiana
-IA: iowa
-KS: kansas
-KY: kentucky
-LA: louisiana
-ME: maine
-MD: maryland
-MA: massachusetts
-MI: michigan
-MN: minnesota
-MS: mississippi
-MO: missouri
-MT: montana
-NE: nebraska
-NV: nevada
-NH: new_hampshire
-NJ: new_jersey
-NM: new_mexico
-NY: new_york
-NC: north_carolina
-ND: north_dakota
-OH: ohio
-OK: oklahoma
-OR: oregon
-PA: pennsylvania
-RI: rhode_island
-SC: south_carolina
-SD: south_dakota
-TN: tennessee
-TX: texas
-UT: utah
-VT: vermont
-VA: virginia
-WA: washington
-WV: west_virginia
-WI: wisconsin
-WY: wyoming
