+++ cerf/data/README.md
@@ -1,1 +1,8 @@
-Space holder for package data.
+Space holder for package data.  Install the cerf package data by running:
+
+```python 
+import cerf
+
+cerf.install_package_data()
+
+```
