+++ MANIFEST.in
@@ -1,3 +1,3 @@
-include cerf/data/*
+include cerf/data/README.md
 include tests/data/*
 include tests/data/comp_data/*
