+++ docs/source/getting_started.rst
@@ -33,6 +33,20 @@ Conda/Miniconda users can utilize the ``environment.yml`` stored in the root of 
 It may be favorable to the user to create a virtual environment for the **cerf** package to minimize package version conflicts.  See `creating virtual environments <https://docs.python.org/3/library/venv.html>`_ to learn how these function and can be setup.
 
 
+Install package data
+--------------------
+
+**cerf** requires package data to be installed from Zenodo to keep the package lightweight.  After **cerf** has been installed, run the following from a Python prompt:
+
+.. code-block:: python
+
+    import cerf
+
+    cerf.install_package_data()
+
+This will automatically download and install the package data necessary to run the examples in accordance with the version of **cerf** you are running.
+
+
 Dependencies
 ------------
 
