### 2021-08-18 #3601 @crvernon
@willu47 I have just merged functionality to incorporate your recent suggestion to separate the data (~200 MB) from the package.  The package data can now be installed via a simple call which is all detailed in the docs via:

```python
import cerf

# install package data
cerf.install_package_data()
```

or from the terminal after installing `cerf` via:

```bash
python -c 'import cerf; cerf.install_package_data()'
```

Thanks for the suggestion!

Please let me know if you have any more questions. 👍
