### 2021-08-19 IMMM-SFA/cerf#57 @willu47
JOSS Review: Big picture
Part of [#3601](https://github.com/openjournals/joss-reviews/issues/3601)

@crvernon - As I've dug into the package, I've seen a lot of useful functionality which could benefit other researchers and would make cerf a more widely used package.

The core of what cerf provides, the spatial allocation of generation plants given local marginal prices and feasibility constraints, is something that could see broader uptake of the package in domains outside of CONUS.

During the review, we've "discussed" indirectly the scope of the package, and its purpose. I'm still a little unclear as to whether:

1. cerf is focussed (limited?) on conducting this analysis for CONUS, in which case it provides data and a collection of classes and methods that implement this, but as a software package is of limited immediate general interest for users outside of CONUS.
2. cerf is a package which implements a particular approach to spatial allocation for energy plants, providing a suite of software components that can be applied in different jurisdictions to this topic of spatial allocation of energy plants, and with specific data requirements.

Given the already relatively niche attraction of the package functionality, I believe that if 1. is the case then the package is not of sufficiently broad interest for the JOSS audience. Whereas with 2. I can see a clear case for publication in JOSS.  But perhaps I am wrong here?

From a development perspective, we've tackled the data issue already (#52), and I think moving from 1. to 2. would require a concerted, but relatively small effort - mainly tying up loose ends (such as #56), especially as the documentation of data requirements etc. are already very good.

I'm happy to discuss this below and it would be good to hear your thoughts? Perhaps it is also worth asking the editor (@fraukewiese) for their input and advice on the above?
