### 2021-08-19 IMMM-SFA/cerf#57 @crvernon
Thanks for the inquiry @willu47 .

Though `cerf` is demonstrated for the CONUS, the package could easily be used in research ranging from regional to global analysis.

`cerf` requires the following inputs to be able to operate:

- A shapefile of substations as points
- A shapefile of gas pipelines  as polylines if siting gas technologies
- A raster containing region ids.  Each grid cell has the value of its parent region.  Currently, this is demonstrated using US  states, though this could be generalized to any region/country/locale identifier matching what the electricity expansion plan provides.
- A raster containing locational marginal pricing (LMP) zone ids.  Each grid cell has the ID of its LMP zones.  These zones are representative of the nodal structure of the underlying model or software that produces the LMPs.  Production cost models to generate this value are not US-specific and this could be generated from any number of open-source and/or commercial products.
- A file of LMPs per zone (corresponding to the aforementioned raster) for each 8760 hour.  Again, the source model or dataset that these are generated from does not have to be US-specific.
- An electricity capacity expansion plan and its accompanying technology-specific information (e.g., variable O&M, fuel price, etc.).  For the CONUS, we use the 50 US state version of GCAM.  However, GCAM is a global model and can produce this data for each of its regions.
- Both the `state-abbrev_to_state-name.yml` and the `state-name_to_state-id.yml` files can simply be replaced with regional information instead.
- Other variable information which is able to be modified per run, are the assumed interconnection costs per km to a substation KV class and the cost per km to connect to a gas pipeline if applicable.
- Lastly, are the suitability rasters which can be composed of any number of locally-relevant exclusion criteria (e.g., protected area, critical habitat, etc.) per technology being sited in the expansion plan.

So `cerf` is generalizable to regions other than the CONUS and is only constrained by the input data that it has been given.  In fact, I would quite like to see `cerf` used in non-US applications.  I conduct research on both US and global fronts and agree that creating generalizable modeling software is important.

I propose doing the following to ameliorate this generalizability:

- rename `state` to `region` where applicable in the inputs and codebase as you had earlier suggested
- expose any CONUS-specific data that is being used in `cerf` to be configurable in the configuration YAML file.  I believe this should only be one or two files but I'll make sure of that
- clearly state in the docs a section on generalizability and how to do so

To this end, I believe `cerf` to be well-suited for JOSS.  Thanks for the input!
