### 2021-08-20 IMMM-SFA/cerf#57 @willu47
Hi @crvernon. Great, this was the answer I was hoping for.  Let me know when you're done with implementing this generalised version and I'll take another look.
