### 2021-08-23 IMMM-SFA/cerf#58 @lisazeyen
JOSS Review: Installation
Part of [#3601](https://github.com/openjournals/joss-reviews/issues/3601)

The installation worked as outlined in the documentation.

-  The package size is very large. It would be useful for a user to have a badge of the repository size in the readme and also in the documentation.

- Some of the data is already loaded by an extra function. Is it possible to load more data (e.g. the folder [comp_data](https://github.com/IMMM-SFA/cerf/tree/main/tests/data/comp_data)) via a function instead of saving it in the repository or reduce the data size?
