### 2021-08-23 IMMM-SFA/cerf#59 @lisazeyen
BUG: JOSS Review: Quickstarter
Part of [#3601](https://github.com/openjournals/joss-reviews/issues/3601)

- [x] I have checked that this issue has not already been reported.

- [x] I have confirmed this bug exists on the latest version of cerf.

- [x] (optional) I have confirmed this bug exists on the master branch of cerf.

---

**Note**: Please read [this guide](https://matthewrocklin.com/blog/work/2018/02/28/minimal-bug-reports) detailing how to provide the necessary information for us to reproduce your bug.

#### Code Sample, a copy-pastable example

```python
# run the configuration for the target year and return a data frame
result_df = cerf.run(config_file, write_output=False)
```

#### Problem description

When running the quickstarter jupyter notebook I get an error in line 3
#### Error message

```python
2021-08-23 09:48:27,147 - root - INFO - Starting CERF model
2021-08-23 09:48:27,322 - root - INFO - Staging data...
2021-08-23 09:48:27,430 - root - INFO - Using 'zones_raster_file':  /home/ws/bw0928/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/data/lmp_zones_1km.img
2021-08-23 09:48:27,473 - root - INFO - Processing locational marginal pricing (LMP)
2021-08-23 09:48:27,474 - root - INFO - Using LMP from default illustrative package data:  /home/ws/bw0928/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/data/illustrative_lmp_8760-per-zone_dollars-per-mwh.zip
2021-08-23 09:48:39,617 - root - INFO - Calculating interconnection costs (IC)
2021-08-23 09:48:39,618 - root - INFO - Using default substations costs from file: /home/ws/bw0928/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/data/costs_per_kv_substation.yml
2021-08-23 09:48:39,621 - root - INFO - Using default substation file: /home/ws/bw0928/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/data/hifld_substations_conus_albers.zip
CPLE_OpenFailedError                      Traceback (most recent call last)
rasterio/_base.pyx in rasterio._base.DatasetBase.__init__()

rasterio/_shim.pyx in rasterio._shim.open_dataset()

rasterio/_err.pyx in rasterio._err.exc_wrap_pointer()

CPLE_OpenFailedError: /tmp/tmpv9o5nkwa/cerf_transmission_distance_substations.tif: No such file or directory

RasterioIOError                           Traceback (most recent call last)
/tmp/ipykernel_32324/2829821130.py in <module>
      6
      7 # run the configuration for the target year and return a data frame
----> 8 result_df = cerf.run(config_file, write_output=False)

~/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/process.py in run(config_file, config_dict, write_output, n_jobs, method, initialize_site_data, log_level)
    181
    182         # process supporting data
--> 183         data = model.stage()
    184
    185         # process all CERF CONUS states in parallel and store the result as a 2D arrays containing sites as

~/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/model.py in stage(self)
     71
     72         # prepare all data for state level run
---> 73         data = Stage(self.settings_dict,
     74                      self.lmp_zone_dict,
     75                      self.technology_dict,

~/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/stage.py in __init__(self, settings_dict, lmp_zone_dict, technology_dict, technology_order, infrastructure_dict, initialize_site_data)
     93         # get interconnection cost per tech [tech_order, x, y]
     94         logging.info('Calculating interconnection costs (IC)')
---> 95         self.ic_arr = self.calculate_ic()
     96
     97         # get NOV array per tech [tech_order, x, y]

~/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/stage.py in calculate_ic(self)
    151
    152         # instantiate class
--> 153         ic = Interconnection(template_array=self.lmp_arr,
    154                              technology_dict=self.technology_dict,
    155                              technology_order=self.technology_order,

~/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/interconnect.py in __init__(self, template_array, technology_dict, technology_order, substation_file, transmission_costs_dict, transmission_costs_file, pipeline_costs_dict, pipeline_costs_file, pipeline_file, output_rasterized_file, output_dist_file, output_alloc_file, output_cost_file, interconnection_cost_file, output_dir)
    119
    120         # calculate electricity transmission infrastructure costs
--> 121         self.substation_costs = self.transmission_to_cost_raster(setting='substations')
    122
    123         # if there are any gas technlogies present, calculate gas pipeline infrastructure costs

~/miniconda3/envs/cerf/lib/python3.9/site-packages/cerf/interconnect.py in transmission_to_cost_raster(self, setting)
    331             alloc_result = wbt.euclidean_allocation(out_rast, out_alloc, callback=suppress_callback)
    332
--> 333             with rasterio.open(out_dist) as dist:
    334                 dist_arr = dist.read(1)
    335

~/miniconda3/envs/cerf/lib/python3.9/site-packages/rasterio/env.py in wrapper(*args, **kwds)
    433
    434         with env_ctor(session=session):
--> 435             return f(*args, **kwds)
    436
    437     return wrapper

~/miniconda3/envs/cerf/lib/python3.9/site-packages/rasterio/__init__.py in open(fp, mode, driver, width, height, count, crs, transform, dtype, nodata, sharing, **kwargs)
    218         # None.
    219         if mode == 'r':
--> 220             s = DatasetReader(path, driver=driver, sharing=sharing, **kwargs)
    221         elif mode == "r+":
    222             s = get_writer_for_path(path, driver=driver)(

rasterio/_base.pyx in rasterio._base.DatasetBase.__init__()

RasterioIOError: /tmp/tmpv9o5nkwa/cerf_transmission_distance_substations.tif: No such file or directory
```
#### Conda environment
```python
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                 conda_forge    conda-forge
_openmp_mutex             4.5                       1_gnu    conda-forge
affine                    2.3.0                    pypi_0    pypi
attrs                     21.2.0                   pypi_0    pypi
backcall                  0.2.0              pyh9f0ad1d_0    conda-forge
backports                 1.0                        py_2    conda-forge
backports.functools_lru_cache 1.6.4              pyhd8ed1ab_0    conda-forge
ca-certificates           2021.7.5             h06a4308_1
cerf                      2.0.3                    pypi_0    pypi
certifi                   2021.5.30                pypi_0    pypi
charset-normalizer        2.0.4                    pypi_0    pypi
click                     8.0.1                    pypi_0    pypi
click-plugins             1.1.1                    pypi_0    pypi
cligj                     0.7.2                    pypi_0    pypi
cloudpickle               1.6.0                      py_0    conda-forge
cycler                    0.10.0                   pypi_0    pypi
debugpy                   1.4.1            py39he80948d_0    conda-forge
decorator                 5.0.9              pyhd8ed1ab_0    conda-forge
fiona                     1.8.20                   pypi_0    pypi
geopandas                 0.9.0                    pypi_0    pypi
idna                      3.2                      pypi_0    pypi
ipykernel                 6.2.0            py39hef51801_0    conda-forge
ipython                   7.26.0           py39hef51801_0    conda-forge
ipython_genutils          0.2.0                      py_1    conda-forge
jedi                      0.18.0           py39hf3d152e_2    conda-forge
joblib                    1.0.1                    pypi_0    pypi
jupyter_client            6.1.12             pyhd8ed1ab_0    conda-forge
jupyter_core              4.7.1            py39hf3d152e_0    conda-forge
kiwisolver                1.3.1                    pypi_0    pypi
ld_impl_linux-64          2.36.1               hea4e1c9_2    conda-forge
libffi                    3.3                  h58526e2_2    conda-forge
libgcc-ng                 11.1.0               hc902ee8_8    conda-forge
libgomp                   11.1.0               hc902ee8_8    conda-forge
libsodium                 1.0.18               h516909a_1    conda-forge
libstdcxx-ng              11.1.0               h56837e0_8    conda-forge
matplotlib                3.4.3                    pypi_0    pypi
matplotlib-inline         0.1.2              pyhd8ed1ab_2    conda-forge
munch                     2.5.0                    pypi_0    pypi
ncurses                   6.2                  h58526e2_4    conda-forge
numpy                     1.21.2                   pypi_0    pypi
openssl                   1.1.1k               h7f98852_1    conda-forge
pandas                    1.3.2                    pypi_0    pypi
parso                     0.8.2              pyhd8ed1ab_0    conda-forge
pexpect                   4.8.0              pyh9f0ad1d_2    conda-forge
pickleshare               0.7.5           py39hde42818_1002    conda-forge
pillow                    8.3.1                    pypi_0    pypi
pip                       21.2.4             pyhd8ed1ab_0    conda-forge
prompt-toolkit            3.0.19             pyha770c72_0    conda-forge
ptyprocess                0.7.0              pyhd3deb0d_0    conda-forge
pygments                  2.10.0             pyhd8ed1ab_0    conda-forge
pyparsing                 2.4.7                    pypi_0    pypi
pyproj                    3.1.0                    pypi_0    pypi
python                    3.9.6           h49503c6_1_cpython    conda-forge
python-dateutil           2.8.2              pyhd8ed1ab_0    conda-forge
python_abi                3.9                      2_cp39    conda-forge
pytz                      2021.1                   pypi_0    pypi
pyyaml                    5.4.1                    pypi_0    pypi
pyzmq                     22.2.1           py39h37b5a0c_0    conda-forge
rasterio                  1.2.6                    pypi_0    pypi
readline                  8.1                  h46c0cb4_0    conda-forge
requests                  2.26.0                   pypi_0    pypi
rtree                     0.9.7                    pypi_0    pypi
scipy                     1.7.1                    pypi_0    pypi
seaborn                   0.11.2                   pypi_0    pypi
setuptools                57.4.0           py39hf3d152e_0    conda-forge
shapely                   1.7.1                    pypi_0    pypi
six                       1.16.0             pyh6c4a22f_0    conda-forge
snuggs                    1.4.7                    pypi_0    pypi
spyder-kernels            2.1.0            py39hf3d152e_0    conda-forge
sqlite                    3.36.0               h9cd32fc_0    conda-forge
tk                        8.6.11               h21135ba_0    conda-forge
tornado                   6.1              py39h3811e60_1    conda-forge
traitlets                 5.0.5                      py_0    conda-forge
tzdata                    2021a                he74cb21_1    conda-forge
urllib3                   1.26.6                   pypi_0    pypi
wcwidth                   0.2.5              pyh9f0ad1d_2    conda-forge
wheel                     0.37.0             pyhd8ed1ab_1    conda-forge
whitebox                  1.5.2                    pypi_0    pypi
wurlitzer                 3.0.0            py39hf3d152e_0    conda-forge
xarray                    0.19.0                   pypi_0    pypi
xz                        5.2.5                h516909a_1    conda-forge
zeromq                    4.3.4                h9c3ff4c_0    conda-forge
zlib                      1.2.11            h516909a_1010    conda-forge
```
