### 2021-08-23 9b2d5dc
update output fn to use input boundary
