+++ cerf/interconnect.py
@@ -1,6 +1,5 @@
 import os
 import logging
-import pkg_resources
 import tempfile
 
 import rasterio
@@ -10,10 +9,11 @@ import yaml
 import numpy as np
 import geopandas as gpd
 
+import cerf.package_data as pkg
+
 from rasterio import features
 
 from cerf.utils import suppress_callback
-from cerf.package_data import cerf_crs, costs_per_kv_substation, get_region_abbrev_to_name
 
 # instantiate whitebox toolset
 wbt = whitebox.WhiteboxTools()
@@ -143,7 +143,7 @@ class Interconnection:
             logging.info(f"Using gas pipeline costs from file:  {f}")
 
         else:
-            f = pkg_resources.resource_filename('cerf', 'data/costs_gas_pipeline.yml')
+            f = pkg.get_costs_gas_pipeline()
             logging.info(f"Using gas pipeline costs from default file:  {f}")
 
         with open(f, 'r') as yml:
@@ -151,35 +151,27 @@ class Interconnection:
 
         return yaml_dict.get('gas_pipeline_cost')
 
-    def process_hifld_substations(self):
-        """Select substations from HIFLD data that are within the CONUS and either in service or under construction and
-        having a minimum voltage rating >= 0.  A field used to rasterize ('_rval_') is also added containing the cost of
-        connection in $/km for each substation.
-
-        This data is assumed to have the following fields:  ['TYPE', 'STATE', 'STATUS'].
-
-        :returns:                               A geodataframe containing the target substations
-
-        """
+    def process_substations(self):
+        """Process input substations from shapefile."""
 
         # load cost dictionary from package data if none passed
         if (self.transmission_costs_dict is None) and (self.transmission_costs_file is None):
-            default_kv_file = pkg_resources.resource_filename('cerf', 'data/costs_per_kv_substation.yml')
+            default_kv_file = pkg.get_costs_per_kv_substation_file()
 
-            logging.info(f"Using default substations costs from file: {default_kv_file}")
+            logging.info(f"Using default substation costs from file: {default_kv_file}")
 
-            self.transmission_costs_dict = costs_per_kv_substation()
+            self.transmission_costs_dict = pkg.costs_per_kv_substation()
 
         elif self.transmission_costs_file is not None:
 
-            logging.info(f"Using substations costs from file: {self.transmission_costs_file}")
+            logging.info(f"Using substation costs from file: {self.transmission_costs_file}")
 
             with open(self.transmission_costs_file, 'r') as yml:
                 self.transmission_costs_dict = yaml.load(yml, Loader=yaml.FullLoader)
 
         if self.substation_file is None:
 
-            sub_file = pkg_resources.resource_filename('cerf', 'data/hifld_substations_conus_albers.zip')
+            sub_file = pkg.get_substation_file()
 
             logging.info(f"Using default substation file: {sub_file}")
 
@@ -190,25 +182,20 @@ class Interconnection:
             logging.info(f"Using substation file: {self.substation_file}")
 
             # get region abbreviations file from cerf package data
-            regions = get_region_abbrev_to_name()
-
-            # load cerf's default coordinate reference system object
-            target_crs = cerf_crs()
+            regions = pkg.get_region_abbrev_to_name()
 
             # load and reproject
-            gdf = gpd.read_file(self.substation_file).to_crs(target_crs)
+            gdf = gpd.read_file(self.substation_file)
 
-            # keep only substations in the CONUS that are either in service or under construction
-            gdf = gdf.loc[(gdf['TYPE'] == 'SUBSTATION') &
-                          (gdf['STATE'].isin(states.keys())) &
[…]
