+++ source/modules.rst
@@ -1,0 +1,9 @@
+cerf
+====
+
+.. toctree::
+   :maxdepth: 4
+
+   cerf
+   setup
+   tests

[…]
