### 2021-08-23 d535c9e
set default settings paths
