+++ cerf/interconnect.py
@@ -39,6 +39,18 @@ class Interconnection:
     :param technology_order:                Order of technologies to process
     :type technology_order:                 list
 
+    :param region_raster_file:              Full path with file name and extension to the region raster file that
+                                            assigns a region ID to each raster grid cell
+    :type region_raster_file:               str
+
+    :param region_abbrev_to_name_file:      Full path with file name and extension to the region abbreviation to name
+                                            YAML reference file
+    :type region_abbrev_to_name_file:       str
+
+    :param region_name_to_id_file:          Full path with file name and extension to the region name to ID YAML
+                                            reference file
+    :type region_name_to_id_file:           str
+
     :param substation_file:                 Full path with file name and extension to the input substations shapefile.
                                             If None, CERF will use the default data stored in the package.
 
@@ -96,7 +108,8 @@ class Interconnection:
 
     """
 
-    def __init__(self, template_array, technology_dict, technology_order, substation_file=None,
+    def __init__(self, template_array, technology_dict, technology_order, region_raster_file,
+                 region_abbrev_to_name_file, region_name_to_id_file, substation_file=None,
                  transmission_costs_dict=None, transmission_costs_file=None, pipeline_costs_dict=None, 
                  pipeline_costs_file=None,pipeline_file=None, output_rasterized_file=False, output_dist_file=False, 
                  output_alloc_file=False, output_cost_file=False, interconnection_cost_file=None, output_dir=None):
@@ -104,6 +117,9 @@ class Interconnection:
         self.template_array = template_array
         self.technology_dict = technology_dict
         self.technology_order = technology_order
+        self.region_raster_file = region_raster_file
+        self.region_abbrev_to_name_file = region_abbrev_to_name_file
+        self.region_name_to_id_file = region_name_to_id_file
         self.substation_file = substation_file
         self.transmission_costs_dict = transmission_costs_dict
         self.transmission_costs_file = transmission_costs_file
@@ -181,9 +197,6 @@ class Interconnection:
 
             logging.info(f"Using substation file: {self.substation_file}")
 
-            # get region abbreviations file from cerf package data
-            regions = pkg.get_region_abbrev_to_name()
-
             # load and reproject
             gdf = gpd.read_file(self.substation_file)
 
@@ -201,8 +214,8 @@ class Interconnection:
 
             return gdf
 
-    def process_eia_natural_gas_pipelines(self):
-        """Select natural gas pipelines from EIA data that have a status of operating and a length greater than 0.
+    def process_pipelines(self):
+        """Select natural gas pipelines data that have a length greater than 0.
 
         :returns:                               A geodataframe containing the target pipelines
 
@@ -226,17 +239,11 @@ class Interconnection:
 
             logging.info(f"Using gas pipeline file:  {self.pipeline_file}")
 
-            # load cerf's default coordinate reference system object
-            target_crs = pkg.cerf_crs()
-
             # read in data and reproject
-            gdf = gpd.read_file(self.pipeline_file).to_crs(target_crs)
+            gdf = gpd.read_file(self.pipeline_file)
 
             # only keep features with a length > 0
             gdf = gdf.loc[gdf.geometry.length > 0].copy()
-
-            # only keep operational pipelines
-            gdf = gdf.loc[gdf['Status'] == 'Operating'].copy()
 
             # set field for rasterize
             gdf['_rval_'] = self.get_pipeline_costs()
@@ -260,15 +267,12 @@ class Interconnection:
             infrastructure_gdf = self.process_substations()
 
         elif setting == 'pipelines':
[…]
