+++ cerf/process_region.py
@@ -140,6 +140,7 @@ class ProcessRegion:
 
         # load the region raster as array
         region_raster_file = self.settings_dict.get('region_raster_file')
+
         with rasterio.open(region_raster_file) as src:
             regions_arr = src.read(1)
 
