+++ cerf/model.py
@@ -88,24 +88,24 @@ class Model(ReadConfig):
         data = self.stage()
 
         process = process_region(target_region_name=target_region_name,
-                                settings_dict=self.settings_dict,
-                                technology_dict=self.technology_dict,
-                                technology_order=self.technology_order,
-                                expansion_dict=self.expansion_dict,
-                                regions_dict=self.regions_dict,
-                                suitability_arr=data.suitability_arr,
-                                lmp_arr=data.lmp_arr,
-                                nov_arr=data.nov_arr,
-                                ic_arr=data.ic_arr,
-                                nlc_arr=data.nlc_arr,
-                                zones_arr=data.zones_arr,
-                                xcoords=data.xcoords,
-                                ycoords=data.ycoords,
-                                indices_2d=data.indices_2d,
-                                randomize=self.settings_dict.get('randomize', True),
-                                seed_value=self.settings_dict.get('seed_value', 0),
-                                verbose=self.settings_dict.get('verbose', False),
-                                write_output=write_output)
+                                 settings_dict=self.settings_dict,
+                                 technology_dict=self.technology_dict,
+                                 technology_order=self.technology_order,
+                                 expansion_dict=self.expansion_dict,
+                                 regions_dict=self.regions_dict,
+                                 suitability_arr=data.suitability_arr,
+                                 lmp_arr=data.lmp_arr,
+                                 nov_arr=data.nov_arr,
+                                 ic_arr=data.ic_arr,
+                                 nlc_arr=data.nlc_arr,
+                                 zones_arr=data.zones_arr,
+                                 xcoords=data.xcoords,
+                                 ycoords=data.ycoords,
+                                 indices_2d=data.indices_2d,
+                                 randomize=self.settings_dict.get('randomize', True),
+                                 seed_value=self.settings_dict.get('seed_value', 0),
+                                 verbose=self.settings_dict.get('verbose', False),
+                                 write_output=write_output)
 
         logging.info(f"CERF model run completed in {round(time.time() - self.start_time, 7)} seconds")
 
