+++ cerf/read_config.py
@@ -3,6 +3,7 @@ import os
 
 import yaml
 
+import cerf.utils as utils
 import cerf.package_data as pkg
 from cerf.logger import Logger
 
@@ -60,9 +61,6 @@ class ReadConfig(Logger):
         self.infrastructure_dict = self.config.get('infrastructure', {})
         self.infrastructure_dict.update(config_dict.get('infrastructure', {}))
 
-        # get the regions dictionary
-        self.regions_dict = self.get_regions_dict()
-
         # update the settings paths to the package defaults if settings are not declared
         region_raster_file = self.settings_dict.get('region_raster_file', None)
         region_abbrev_to_name_file = self.settings_dict.get('region_abbrev_to_name_file', None)
@@ -72,10 +70,19 @@ class ReadConfig(Logger):
             self.settings_dict.update({'region_raster_file': pkg.cerf_regions_raster()})
 
         if region_abbrev_to_name_file is None:
-            self.settings_dict.update({'region_abbrev_to_name_file': pkg.get_region_abbrev_to_name()})
+            self.settings_dict.update({'region_abbrev_to_name_file': pkg.get_region_abbrev_to_name_file()})
 
         if region_name_to_id_file is None:
             self.settings_dict.update({'region_name_to_id_file': pkg.get_region_name_to_id()})
+
+        # validate file existence
+        self.validate_settings_files()
+        self.validate_lmp_files()
+        self.validate_technology_files()
+        self.validate_infrastructure_files()
+
+        # get the regions dictionary
+        self.regions_dict = self.get_regions_dict()
 
     @staticmethod
     def read_yaml(yaml_file):
@@ -103,7 +110,9 @@ class ReadConfig(Logger):
             return self.read_yaml(self.config_file)
 
         else:
-            msg = f"Config file not found for path:  {self.config_file}"
+            msg = f"""Config file not found for path:  {self.config_file}. If using defaults, please download the package data. 
+            See https://immm-sfa.github.io/cerf/getting_started.html#install-package-data"""
+
             logging.error(msg)
             raise FileNotFoundError(msg)
 
@@ -115,8 +124,79 @@ class ReadConfig(Logger):
 
         return self.read_yaml(regions_lookup_file)
 
-    # TODO
-    def validate_file_exists(self):
-        """Ensure that files necessary to run the package exists."""
+    def validate_settings_files(self):
+        """Ensure that files necessary files exists for settings."""
 
-        pass
+        # expected files from settings
+        settings_files = [self.settings_dict['region_raster_file'],
+                          self.settings_dict['region_abbrev_to_name_file'],
+                          self.settings_dict['region_name_to_id_file']]
+
+        for i in settings_files:
+            if not os.path.isfile(i):
+                msg = f"""Cannot find the settings file for: {i}.  If using defaults, please download the package data. 
+                See https://immm-sfa.github.io/cerf/getting_started.html#install-package-data"""
+
+                raise FileNotFoundError(msg)
+
+    def validate_technology_files(self):
+        """Ensure that files necessary files exists for technology."""
+
+        # expected files from technology
+        suitability_file_dict = utils.default_suitabiity_files()
+
+        for i in self.technology_dict.keys():
+
+            suitability_raster = self.technology_dict[i]['suitability_raster_file']
+            tech_name = self.technology_dict[i]['tech_name']
+
+            # see if the default suitability file exists in the data directory
+            if suitability_raster is None:
+
+                suit_path = os.path.join(pkg.get_data_directory(), suitability_file_dict[tech_name])
+
+                if not os.path.isfile(suit_path):
+                    msg = f"""Cannot find the default suitability raster: {suit_path}.  Please download the package data. 
[…]
