### 2021-08-24 42ae552
update read config tests
