+++ tests/data/test_config_2010.yml
@@ -15,11 +15,15 @@ settings:
     # if `randomize` is False, set a seed value for reproducibility; the default is 0
     seed_value: 0
 
+    region_raster_file: null
+    region_abbrev_to_name_file: null
+    region_name_to_id_file: null
+
 # technology specific information
 technology:
 
     9:
-        tech_name: biomass
+        tech_name: biomass_conv_wo_ccs
         lifetime: 60.0
         capacity_factor: 0.6090000000000005
         variable_cost_esc_rate: -0.00398993418629034
@@ -39,7 +43,7 @@ technology:
         suitability_raster_file: null
 
     11:
-        tech_name: nuclear
+        tech_name: nuclear_gen_ii
         lifetime: 60.0
         capacity_factor: 0.8999999999999995
         variable_cost_esc_rate: -0.00104311614063357
@@ -63,454 +67,459 @@ expansion_plan:
 
     alabama:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 0
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 4
 
     arizona:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 5
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 0
 
     arkansas:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 10
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 3
 
     california:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 1
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 2
 
     colorado:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 9
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 1
 
     connecticut:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 2
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 9
 
     delaware:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 4
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 9
 
     district_of_columbia:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 9
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 4
 
     florida:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 8
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 4
 
     georgia:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 6
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 3
 
     idaho:
         9:
-            tech_name: biomass
+            tech_name: biomass_conv_wo_ccs
             n_sites: 10
 
         11:
-            tech_name: nuclear
+            tech_name: nuclear_gen_ii
             n_sites: 9
 
     illinois:
         9:
-            tech_name: biomass
[…]
