+++ docs/source/modules.rst
@@ -1,5 +1,5 @@
-cerf
-====
+API documentation
+=================
 
 .. toctree::
    :maxdepth: 4
