+++ tests/data/test_config_2010.yml
@@ -1,525 +1,0 @@
-# project settings
-settings:
-
-    # target year to run
-    run_year: 2010
-
-    # location to store the output data to
-    output_directory: null
-
-    # randomize selection of a site for a technology when NLC values are equal
-    #  the first pass is always random but setting `randomize` to False and
-    #  passing a seed value will ensure that runs are reproducible
-    randomize: True
-
-    # if `randomize` is False, set a seed value for reproducibility; the default is 0
-    seed_value: 0
-
-    region_raster_file: null
-    region_abbrev_to_name_file: null
-    region_name_to_id_file: null
-
-# technology specific information
-technology:
-
-    9:
-        tech_name: biomass_conv_wo_ccs
-        lifetime: 60.0
-        capacity_factor: 0.6090000000000005
-        variable_cost_esc_rate: -0.00398993418629034
-        fuel_esc_rate: 0.0
-        unit_size: 80
-        interconnection_cost_per_km: 552
-        variable_om: 11.68495803744351
-        heat_rate: 15117.64999999997
-        fuel_price: 0.0
-        carbon_capture_rate: 0.0
-        fuel_co2_content: 0.3035999999999996
-        discount_rate: 0.05
-        carbon_esc_rate: 0.0
-        carbon_tax: 0.0
-        buffer_in_km: 5
-        require_pipelines: False
-        suitability_raster_file: null
-
-    11:
-        tech_name: nuclear_gen_ii
-        lifetime: 60.0
-        capacity_factor: 0.8999999999999995
-        variable_cost_esc_rate: -0.00104311614063357
-        fuel_esc_rate: 0.04639000000000001
-        unit_size: 1350
-        interconnection_cost_per_km: 1104
-        variable_om: 2.098127824402843
-        heat_rate: 10246.19999999998
-        fuel_price: 0.7128099999999996
-        carbon_capture_rate: 0.0
-        fuel_co2_content: 0.0
-        discount_rate: 0.05
-        carbon_esc_rate: 0.0
-        carbon_tax: 0.0
-        buffer_in_km: 5
-        require_pipelines: False
-        suitability_raster_file: null
-
-# expansion plan from GCAM
-expansion_plan:
-
-    alabama:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 0
-
-        11:
-            tech_name: nuclear_gen_ii
-            n_sites: 4
-
-    arizona:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 5
-
-        11:
-            tech_name: nuclear_gen_ii
-            n_sites: 0
-
-    arkansas:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 10
-
-        11:
-            tech_name: nuclear_gen_ii
-            n_sites: 3
-
-    california:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 1
-
-        11:
-            tech_name: nuclear_gen_ii
-            n_sites: 2
-
-    colorado:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 9
-
-        11:
-            tech_name: nuclear_gen_ii
-            n_sites: 1
-
-    connecticut:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 2
-
-        11:
-            tech_name: nuclear_gen_ii
-            n_sites: 9
-
-    delaware:
-        9:
-            tech_name: biomass_conv_wo_ccs
-            n_sites: 4
-
-        11:
-            tech_name: nuclear_gen_ii
[…]
