+++ cerf/interconnect.py
@@ -288,53 +288,53 @@ class Interconnection:
             # get shapes
             shapes = ((geom, value) for geom, value in zip(infrastructure_gdf.geometry, infrastructure_gdf['_rval_']))
 
-        with tempfile.TemporaryDirectory() as tempdir:
+        tempdir = tempfile.gettempdir()
 
-            # if write desired
-            if any((self.output_rasterized_file, self.output_dist_file, self.output_alloc_file, self.output_cost_file)):
+        # if write desired
+        if any((self.output_rasterized_file, self.output_dist_file, self.output_alloc_file, self.output_cost_file)):
 
-                if self.output_dir is None:
-                    msg = "If writing rasters to file must specify 'output_dir'"
-                    logging.error(msg)
-                    raise NotADirectoryError(msg)
+            if self.output_dir is None:
+                msg = "If writing rasters to file must specify 'output_dir'"
+                logging.error(msg)
+                raise NotADirectoryError(msg)
 
-                else:
-                    out_rast = os.path.join(self.output_dir, f'cerf_transmission_raster_{setting}.tif')
-                    out_dist = os.path.join(self.output_dir, f'cerf_transmission_distance_{setting}.tif')
-                    out_alloc = os.path.join(self.output_dir, f'cerf_transmission_allocation_{setting}.tif')
-                    out_costs = os.path.join(self.output_dir, f'cerf_transmission_costs_{setting}.tif')
             else:
-                out_rast = os.path.join(tempdir, f'cerf_transmission_raster_{setting}.tif')
-                out_dist = os.path.join(tempdir, f'cerf_transmission_distance_{setting}.tif')
-                out_alloc = os.path.join(tempdir, f'cerf_transmission_allocation_{setting}.tif')
-                out_costs = os.path.join(tempdir, f'cerf_transmission_costs_{setting}.tif')
+                out_rast = os.path.join(self.output_dir, f'cerf_transmission_raster_{setting}.tif')
+                out_dist = os.path.join(self.output_dir, f'cerf_transmission_distance_{setting}.tif')
+                out_alloc = os.path.join(self.output_dir, f'cerf_transmission_allocation_{setting}.tif')
+                out_costs = os.path.join(self.output_dir, f'cerf_transmission_costs_{setting}.tif')
+        else:
+            out_rast = os.path.join(tempdir, f'cerf_transmission_raster_{setting}.tif')
+            out_dist = os.path.join(tempdir, f'cerf_transmission_distance_{setting}.tif')
+            out_alloc = os.path.join(tempdir, f'cerf_transmission_allocation_{setting}.tif')
+            out_costs = os.path.join(tempdir, f'cerf_transmission_costs_{setting}.tif')
 
-            # rasterize transmission vector data and write to memory
-            with rasterio.open(out_rast, 'w', **metadata) as dataset:
-                # burn features into raster
-                burned = features.rasterize(shapes=shapes, fill=0, out=arr, transform=dataset.transform)
+        # rasterize transmission vector data and write to memory
+        with rasterio.open(out_rast, 'w', **metadata) as dataset:
+            # burn features into raster
+            burned = features.rasterize(shapes=shapes, fill=0, out=arr, transform=dataset.transform)
 
-                # write the outputs to file
-                dataset.write_band(1, burned)
+            # write the outputs to file
+            dataset.write_band(1, burned)
 
-            # calculate Euclidean distance and write raster; result just stores the return value 0
-            dist_result = wbt.euclidean_distance(out_rast, out_dist, callback=suppress_callback)
+        # calculate Euclidean distance and write raster; result just stores the return value 0
[…]
