+++ .github/workflows/build-ubuntu-18.04.yml
@@ -1,4 +1,4 @@
-name: build
+name: build-ubuntu-18.04
 
 on: [push, pull_request]
 
