### 2021-08-25 c2750ee
run verbose to debug ub18
