### 2021-08-25 IMMM-SFA/cerf#59 @crvernon
@lisazeyen this took a bit of digging because I wasn't getting the same behavior on my Ubuntu 20.04.2 LTS CI server via GitHub actions.  It seems that one of cerf's dependencies that conducts some of the necessary spatial analysis (whitebox) requires the use of the `GLIBC_2.29` (GNU C library).  Unfortunately, an older version (`GLIBC_2.27`) of this is present in Ubuntu 18.04.

Since the GLIBC is a fundamental library in Ubuntu, I don't feel confident about officially recommending that Ubuntu 18 users upgrade their systems to this new `GLIBC_2.29` version.  However, I will provide a note in the documentation under the install protocol that this may be an option for some users but I will officially claim to only support Ubuntu 20.04.2 LTS and above for Ubuntu users.

Thanks so much for finding this and I'll make the change to the docs shortly!
