+++ README.md
@@ -19,6 +19,8 @@
 
 ### Install `cerf`
 
+**NOTE**:  `cerf` is not officially supported for Ubuntu 18 users due to a system dependency (`GLIB_2.29`) required by the `whitebox` package which `cerf` uses to conduct spatial analysis. Ubuntu 18 natively includes `GLIB_2.27`.  It may be possible for Ubuntu 18 users to upgrade to `GLIB_2.29` but this should be done with careful consideration.  Instead, we officially support `cerf` use for Ubuntu users for versions 20.04.2 LTS and greater.
+
 ```bash
 pip install cerf
 ```
