### 2021-08-25 #3601 @crvernon
:wave: @willu47 and @lisazeyen - I have addressed all of your issues and comments.  At this point we should be up-to-date.  All changes have been incorporated on `main` and the docs and PyPi versions updated accordingly.  Thanks so much for the great feedback!
