### 2021-08-30 028c927
remove stray data file
