### 2021-09-01 IMMM-SFA/cerf#66 @lisazeyen
JOSS Review: Community guidelines
Part of [#3601](https://github.com/openjournals/joss-reviews/issues/3601)

One could add a PR template (similar to the issue templates)
