### 2021-09-01 IMMM-SFA/cerf#67 @lisazeyen
JOSS Review: documentation - add release notes
Part of [#3601](https://github.com/openjournals/joss-reviews/issues/3601)

It would be nice to have the release notes in the documentation.
