### 2021-09-02 2eeac2c
add release notes to docs and update version
