+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,0 +1,24 @@
+**Purpose:**
+
+Short description of the purpose of this PR
+
+**New version:**
+
+Then new version number if one is to be issued
+
+**Related:**
+
+Any related pull requests or issues from other repositories
+
+**Resolved:**
+
+Any issues to resolve automatically upon merge.  Use the keyword with the number of the target issue (e.g., `Resolve #59`)
+
+**Incorporated:**
+
+Any pull requests already merged that this one may include.  This is usually the case when feature and bug fix branches have been merged into the `dev` branch then these are staged to be merged into the `main`
+
+
+**Comments:**
+
+Any additional comments.
