### 2021-09-16 IMMM-SFA/cerf#71 @willu47
JOSS REVIEW: Location of downloaded data is not findable when packages is installed
Review openjournals/joss-reviews#3601

- [x] I have checked that this issue has not already been reported.

- [x] I have confirmed this bug exists on the latest version of cerf.

- [x] (optional) I have confirmed this bug exists on the master branch of cerf.

---

**Note**: Please read [this guide](https://matthewrocklin.com/blog/work/2018/02/28/minimal-bug-reports) detailing how to provide the necessary information for us to reproduce your bug.

#### Code Sample, a copy-pastable example

```python
import cerf

cerf.install_package_data()
```

#### Problem description

The current `install_package_data()` function downloads and places data within the directory where the package is installed.

While this is functional (the package can find the data fine), it's not very helpful for a user either to view the example data, or to modify it to experiment with the package.

If I install cerf using pip `pip install cerf`, then cerf is placed in a system-defined directory within the Python environment I have defined. For example:

`/Users/wusher/miniconda3/envs/cerf_rev_2/lib/python3.9/site-packages/cerf/data/`

#### Expected Output

I think it would be more user friendly and transparent if the data was downloaded to a user-defined location, with a path passed to the function.

To run an analysis this would also require a user to provide a path, or configuration file containing the path to the data.

For example:

```python
import cerf

DATA_PATH = 'data'

cerf.Install_package_data(DATA_PATH)

sample_year = 2030

# load the sample configuration file path for the target year
config_file = cerf.config_file(sample_year, DATA_PATH)

# run the configuration for the target year and return a data frame
result_df = cerf.run(config_file, write_output=False)

result_df.head()
```
