### 2021-09-16 ef2a96e
update install to use local dir optionally
