### 2021-09-17 02baba2
update docs for additional data install option
