### 2021-09-17 e395780
update release notes
