+++ docs/source/release_notes.rst
@@ -7,9 +7,10 @@ see the `commit logs <https://github.com/IMMM-SFA/cerf/commits>`_.
 Version 2.0.7
 _____________
 
-Changes made in `PR #68 <https://github.com/IMMM-SFA/cerf/pull/68>`_
+Changes made in `PR #74 <https://github.com/IMMM-SFA/cerf/pull/74>`_
 
 - Provide option to download illustrative data to a directory of the user's choosing
+- Fix bug in the units of interconnection cost
 
 
 Version 2.0.6
