### 2021-09-17 #3601 @crvernon
@willu47 I added in the option to download the illustrative test data to a user-defined location as described in https://github.com/IMMM-SFA/cerf/pull/72

🤝 Thanks for all of your input and the thoughtful review!  I now have addressed all of your inquiries.
