+++ source/cerf.rst
@@ -1,125 +1,0 @@
-cerf package
-============
-
-Submodules
-----------
-
-cerf.compete module
--------------------
-
-.. automodule:: cerf.compete
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.install\_supplement module
--------------------------------
-
-.. automodule:: cerf.install_supplement
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.interconnect module
-------------------------
-
-.. automodule:: cerf.interconnect
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.lmp module
----------------
-
-.. automodule:: cerf.lmp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.logger module
-------------------
-
-.. automodule:: cerf.logger
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.model module
------------------
-
-.. automodule:: cerf.model
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.nov module
----------------
-
-.. automodule:: cerf.nov
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.outputs module
--------------------
-
-.. automodule:: cerf.outputs
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.package\_data module
--------------------------
-
-.. automodule:: cerf.package_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.process module
--------------------
-
-.. automodule:: cerf.process
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.process\_region module
----------------------------
-
-.. automodule:: cerf.process_region
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.read\_config module
-------------------------
-
-.. automodule:: cerf.read_config
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.stage module
------------------
-
-.. automodule:: cerf.stage
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-cerf.utils module
------------------
-
-.. automodule:: cerf.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: cerf
-   :members:
-   :undoc-members:
-   :show-inheritance:
