+++ source/tests.rst
@@ -1,69 +1,0 @@
-tests package
-=============
-
-Submodules
-----------
-
-tests.test\_compete module
---------------------------
-
-.. automodule:: tests.test_compete
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_lmp module
-----------------------
-
-.. automodule:: tests.test_lmp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_nov module
-----------------------
-
-.. automodule:: tests.test_nov
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_package\_data module
---------------------------------
-
-.. automodule:: tests.test_package_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_read\_config module
--------------------------------
-
-.. automodule:: tests.test_read_config
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_spatial\_reader module
-----------------------------------
-
-.. automodule:: tests.test_spatial_reader
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_utils module
-------------------------
-
-.. automodule:: tests.test_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
