+++ source/modules.rst
@@ -1,9 +1,0 @@
-cerf
-====
-
-.. toctree::
-   :maxdepth: 4
-
-   cerf
-   setup
-   tests
