+++ cerf/__init__.py
@@ -8,4 +8,4 @@ from .utils import *
 from .install_supplement import install_package_data
 
 
-__version__ = "2.0.8"
+__version__ = "2.0.9"
