+++ CITATION.cff
@@ -25,8 +25,8 @@ authors:
 - family-names: "Binsted"
   given-names: "Matthew"
   orcid: "https://orcid.org/0000-0002-5177-7253"
-title: "cerf v2.0.7: an open-source geospatial Python package for evaluating and analyzing future electricity technology capacity expansion feasibility."
-version: 2.0.7
-doi: 10.5281/zenodo.5140077
-date-released: 2021-10-16
+title: "cerf: A Python package to evaluate the feasibility and costs of power plant siting for alternative futures"
+version: 2.0.9
+doi: 10.5281/zenodo.1321312
+date-released: 2021-09-24
 url: "https://github.com/IMMM-SFA/cerf"
