### 2021-09-25 #3601 @crvernon
@fraukewiese My release version is `v2.0.9` and my DOI is 10.5281/zenodo.5527334

Thanks!
