### 2021-09-25 #3601 @crvernon
> @crvernon The title in the archive does not exactly match the paper title: alternative instead of alternate.

@fraukewiese yes, I changed the name in the paper as well. The original should have used "alternative" instead. The current is correct. Thank you!
